requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.3
cryptography==41.0.7
psycopg2-binary==2.9.9
